Bimtester: Fixed attributes_psets step files
